backend/requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-jwt-extended
flask-bcrypt
flask-cors
python-dotenv
spacy
scikit-learn
pdfplumber
docx2txt
requests